pyproject.toml
[tool.poetry]
name = "sf800p2mqtt"
version = "1.8.0"
description = "Sniffer for SolarFlow 800 Pro and MQTT publishing."
authors = ["Ixtalo <ixtalo@users.noreply.github.com>"]
license = "AGPL"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
scapy = "^2.6.1"
python-dotenv = "^1.1.1"
paho-mqtt = "^2.1.0"
colorlog = "^6.9.0"
docopt = "^0.6.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
pylint = "*"
pytest = "^8.4.1"
pytest-cov = "^6.2.1"
pytest-sugar = "^1.0.0"
pytest-clarity = "*"
pytest-icdiff = "*"

[tool.poetry.group.lab]
optional = true

[tool.poetry.group.lab.dependencies]
jupyter = "^1.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

## https://flake8.pycqa.org/en/latest/user/configuration.html
## https://github.com/john-hen/Flake8-pyproject
[tool.flake8]
ignore = ["E501"]
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 120
count = true

## http://www.pydocstyle.org/en/stable/usage.html#configuration-files
[tool.pydocstyle]
#match_dir='^[^\.t].+$'
#match_dir = '[^/].+'

## https://docs.pytest.org/en/7.1.x/reference/customize.html#pyproject-toml
[tool.pytest.ini_options]
minversion = "6.0"
testpaths = ["tests",]

## https://coverage.readthedocs.io/en/latest/source.html#source
[tool.coverage.run]
omit = ["tests/*", "__init__.py"]

## https://github.com/hhatto/autopep8?tab=readme-ov-file#configuration
[tool.autopep8]
max_line_length = 120
ignore = "E501,W6"  # or ["E501", "W6"]
in-place = true
recursive = true
aggressive = 3

[tool.pylint]
max-line-length = 120
